+++ R/renderers.R
@@ -87,16 +87,13 @@ gifski_renderer <- function(file = NULL, loop = TRUE, width = NULL, height = NUL
   if (!requireNamespace('gifski', quietly = TRUE)) {
     stop('The gifski package is required to use gifski_renderer', call. = FALSE)
   }
-  if (!requireNamespace('png', quietly = TRUE)) {
-    stop('The png package is required to use gifski_renderer', call. = FALSE)
-  }
   function(frames, fps) {
     if (is.null(file)) file <- tempfile(fileext = '.gif')
     if (!all(grepl('.png$', frames))) {
       stop('gifski only supports png files', call. = FALSE)
     }
     if (is.null(width) || is.null(height)) {
-      dims <- dim(png::readPNG(frames[1], native = TRUE))
+      dims <- png_dim(frames[1])
       height <- height %||% dims[1]
       width <- width %||% dims[2]
     }